fix/remove-domain-files (#42)
