--- src/domain/address_domain.go
@@ -1,13 +1,0 @@
-package domain
-
-import "cij_api/src/model"
-
-type AddressRepo interface {
-	GetAddressById(id int) (model.Address, error)
-	UpsertAddress(address model.Address) (int, error)
-	DeleteAddress(id int) error
-}
-
-type AddressService interface {
-	GetAddressById(id int) (model.Address, error)
-}
