--- src/domain/news_domain.go
@@ -1,11 +1,0 @@
-package domain
-
-import "cij_api/src/model"
-
-type NewsRepo interface {
-	ListNews() ([]model.News, error)
-}
-
-type NewsService interface {
-	ListNews() ([]model.NewsResponse, error)
-}
